.agentignore

+++ internal/tools/files/delete_file.go
@@ -4,6 +4,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
+
+	"github.com/AamindMandragora/pragma/internal/tools"
 )
 
 type DeleteFileTool struct{}
@@ -34,6 +36,9 @@ func (d *DeleteFileTool) Execute(args json.RawMessage) (string, error) {
 	}
 	if err := json.Unmarshal(args, &params); err != nil {
 		return "", err
+	}
+	if tools.IsIgnored(params.Path) {
+		return "", fmt.Errorf("access denied: %s is in .agentignore", params.Path)
 	}
 	err := os.Remove(params.Path)
 	if err != nil {
